:recycle: Move jsonrpc to vm package
