+++ vm/server/src/jsonrpc/StructuredValue.ts
@@ -1,3 +1,3 @@
-import type { PrimitiveType } from "./PrimitiveType";
+import type { PrimitiveType } from "./PrimitiveType.js";
 
 export type StructuredValue = Record<string, PrimitiveType> | ReadonlyArray<PrimitiveType>
