+++ vm/vm/src/actions/contractCall/RunContractCallError.ts
@@ -1,0 +1,2 @@
+// TODO we want to fill this out with everything that could go wrong via effect
+export type RunContractCallError = Error
