+++ vm/vm/src/actions/contractCall/RunContractCallHandlerGeneric.ts
@@ -1,0 +1,12 @@
+import type { Tevm } from '../../tevm.js'
+import type { Abi } from 'abitype'
+import type { RunContractCallAction } from './RunContractCallAction.js'
+import type { RunContractCallResult } from './RunContractCallResult.js'
+
+export type RunContractCallHandlerGeneric = <
+	TAbi extends Abi | readonly unknown[] = Abi,
+	TFunctionName extends string = string,
+>(
+	tevm: Tevm,
+	action: RunContractCallAction<TAbi, TFunctionName>,
+) => Promise<RunContractCallResult<TAbi, TFunctionName>>
