+++ vm/vm/src/actions/contractCall/defaultCaller.js
@@ -1,0 +1,1 @@
+export const defaultCaller = '0x0000000000000000000000000000000000000000'
