+++ vm/vm/src/actions/contractCall/defaultGasLimit.js
@@ -1,0 +1,1 @@
+export const defaultGasLimit = BigInt(0xfffffffffffff)
