+++ vm/vm/src/actions/contractCall/index.js
@@ -1,0 +1,13 @@
+export { defaultCaller } from './defaultCaller.js'
+export { defaultGasLimit } from './defaultGasLimit.js'
+export { runContractCallHandler } from './runContractCallHandler.js'
+/**
+ * @typedef {import("./RunContractCallAction.js").RunContractCallAction} RunContractCallAction
+ */
+export { RunContractCallActionValidator } from './RunContractCallActionValidator.js'
+/**
+ * @typedef {import("./RunContractCallError.js").RunContractCallError}.RunContractCallError
+ */
+/**
+ * @typedef {import("./RunContractCallResult.js").RunContractCallResult}.RunContractCallResult
+ */
