+++ vm/vm/src/actions/putAccount/DEFAULT_BALANCE.js
@@ -1,0 +1,4 @@
+import { parseEther } from 'viem'
+
+export const DEFAULT_BALANCE = parseEther('1000')
+
