+++ vm/vm/src/actions/putAccount/PutAccountAction.ts
@@ -1,0 +1,9 @@
+import type { Address } from 'abitype'
+
+/**
+ * Tevm action to put an account into the vm state
+ */
+export type PutAccountAction = {
+	account: Address
+	balance?: bigint
+}
