+++ vm/vm/src/actions/putAccount/index.ts
@@ -1,0 +1,3 @@
+export { type PutAccountAction } from './PutAccountAction.js'
+export { PutAccountActionValidator } from './PutAccountActionValidator.js'
+export { putAccountHandler } from './putAccountHandler.js'
