+++ vm/vm/src/actions/putContractCode/index.js
@@ -1,0 +1,5 @@
+export { putContractCodeHandler } from './putContractCodeHandler.js'
+/**
+ * @typedef {import('./PutContractCodeAction.js').PutContractCodeAction} PutContractCodeAction
+ */
+export { PutContractCodeActionValidator } from './PutContractCodeActionValidator.js'
