+++ vm/vm/src/actions/runCall/index.js
@@ -1,0 +1,5 @@
+export { CallActionValidator } from './CallActionValidator.js'
+/**
+ * @typedef {import("./RunCallAction.js").RunCallAction} RunCallAction
+ */
+export { runCallHandler } from './runCallHandler.js'
