+++ vm/vm/src/actions/runScript/RunScriptError.ts
@@ -1,0 +1,2 @@
+export type RunScriptError = Error
+
