+++ vm/vm/src/actions/runScript/RunScriptResult.ts
@@ -1,0 +1,8 @@
+import type { Abi } from 'abitype'
+import type { RunContractCallResult } from '../contractCall/RunContractCallResult.js'
+
+export type RunScriptResult<
+	TAbi extends Abi | readonly unknown[] = Abi,
+	TFunctionName extends string = string,
+> = RunContractCallResult<TAbi, TFunctionName>
+
