+++ vm/vm/src/actions/runScript/index.js
@@ -1,0 +1,8 @@
+export { runScriptHandler } from './runScriptHandler.js'
+export { RunScriptActionValidator } from './RunScriptActionValidator.js'
+/**
+ * @typedef {import("./RunScriptAction.js").RunScriptAction} RunScriptAction
+ */
+/**
+ * @typedef {import("./RunScriptResult.js").RunScriptResult}.RunScriptResult
+ */
