+++ vm/vm/src/utils/ZHex.ts
@@ -1,7 +1,7 @@
-import type { Hex } from 'viem'
+import { type Hex } from 'viem'
 import { z } from 'zod'
+import { hexRegex } from './hexRegex.js'
 
-const hexRegex = /^0x[0-9a-fA-F]*$/
 export const ZHex = z.string().transform((value, ctx) => {
 	if (!hexRegex.test(value)) {
 		ctx.addIssue({
