+++ vm/vm/src/utils/hexRegex.ts
@@ -1,0 +1,1 @@
+export const hexRegex = /^0x[0-9a-fA-F]*$/
